new/src/backend/pyproject.toml
[tool.poetry]
name = "atomic-charge-calculator-2-plus"
version = "0.1.0"
description = ""
authors = ["Martin Pilát <you@example.com>"]
readme = "README.md"
packages = [{include = "app"}]
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.4"}
uvicorn = {extras = ["standard"], version = "^0.32.0"}
python-dotenv = "^1.0.1"
aiofiles = "^24.1.0"
pydantic-settings = "^2.6.1"
dependency-injector = "^4.43.0"
uvicorn-worker = "^0.2.0"
gunicorn = "^23.0.0"
sqlalchemy = "^2.0.37"
psycopg2 = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
